# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bosses:
    get:
      operationId: listBosses
      tags: [bosses]
      summary: List available bosses
      responses:
        "200":
          description: Boss list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Boss"
    post:
      operationId: createCustomBoss
      tags: [bosses]
      summary: Create a custom boss (pro feature)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BossInput"
      responses:
        "201":
          description: Created boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users with pagination
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details with player profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (role, isPro, username)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/player:
    patch:
      operationId: adminUpdatePlayer
      tags: [admin]
      summary: Edit player stats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /admin/bosses:
    get:
      operationId: adminListBosses
      tags: [admin]
      summary: List all bosses
      responses:
        "200":
          description: Boss list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Boss"
    post:
      operationId: adminCreateBoss
      tags: [admin]
      summary: Create a boss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBossInput"
      responses:
        "201":
          description: Created boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
  /admin/bosses/{bossId}:
    patch:
      operationId: adminUpdateBoss
      tags: [admin]
      summary: Update a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBossInput"
      responses:
        "200":
          description: Updated boss
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boss"
    delete:
      operationId: adminDeleteBoss
      tags: [admin]
      summary: Delete a boss
      parameters:
        - name: bossId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/skills:
    get:
      operationId: adminListSkills
      tags: [admin]
      summary: List all skill definitions
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillDefinition"
    post:
      operationId: adminCreateSkill
      tags: [admin]
      summary: Create a skill definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillDefinitionInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDefinition"
  /admin/skills/{skillId}:
    patch:
      operationId: adminUpdateSkill
      tags: [admin]
      summary: Update a skill definition
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillDefinitionInput"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDefinition"
    delete:
      operationId: adminDeleteSkill
      tags: [admin]
      summary: Delete a skill definition
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/items:
    get:
      operationId: adminListItems
      tags: [admin]
      summary: List all inventory items
      parameters:
        - name: rarity
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated item list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminItemList"
    post:
      operationId: adminCreateItem
      tags: [admin]
      summary: Grant an item to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /admin/items/{itemId}:
    delete:
      operationId: adminDeleteItem
      tags: [admin]
      summary: Delete an inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    Player:
      type: object
      required: [id, userId, username, level, xp, xpToNextLevel, hp, maxHp, gold, strength, endurance, dexterity, faith, avatarClass, equippedSkills, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        gold:
          type: integer
        strength:
          type: integer
        endurance:
          type: integer
        dexterity:
          type: integer
        faith:
          type: integer
        avatarClass:
          type: string
        equippedSkills:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Boss:
      type: object
      required: [id, name, description, difficulty, maxHp, xpReward, goldReward, lootTable, imageUrl, isCustom, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
        lootTable:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
    BossInput:
      type: object
      required: [name, description, difficulty, maxHp]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
...
    InventoryItem:
      type: object
      required: [id, userId, name, description, rarity, type, obtainedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        type:
          type: string
          enum: [weapon, armor, accessory, consumable, trophy]
        obtainedAt:
          type: string

    # SKILLS
...
    ActivityEntry:
      type: object
      required: [id, type, description, xpGained, timestamp]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        description:
          type: string
        xpGained:
          type: integer
        timestamp:
          type: string

    # LEADERBOARD
...
    SkillDefinition:
      type: object
      required: [id, name, description, effect, levelRequired]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        effect:
          type: string
        levelRequired:
          type: integer
    SkillDefinitionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        effect:
          type: string
        levelRequired:
          type: integer
    AdminUser:
      type: object
      required: [id, username, role, isPro, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [user, admin, moderator]
        isPro:
          type: boolean
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [user, admin, moderator]
        isPro:
          type: boolean
    AdminPlayerUpdate:
      type: object
      properties:
        level:
          type: integer
        xp:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        gold:
          type: integer
        strength:
          type: integer
        endurance:
          type: integer
        dexterity:
          type: integer
        faith:
          type: integer
        avatarClass:
          type: string
        equippedSkills:
          type: array
          items:
            type: string
    AdminBossInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        maxHp:
          type: integer
        xpReward:
          type: integer
        goldReward:
          type: integer
        lootTable:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
    AdminItemInput:
      type: object
      required: [userId, name]
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        type:
          type: string
          enum: [weapon, armor, accessory, consumable, trophy]
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    AdminUserListEntry:
      type: object
      required: [id, username, role, isPro, createdAt, level, xp, gold, avatarClass]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [user, admin, moderator]
        isPro:
          type: boolean
        createdAt:
          type: string
        level:
          type: integer
        xp:
          type: integer
        gold:
          type: integer
        avatarClass:
          type: string
    AdminUserList:
      type: object
      required: [users, pagination]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserListEntry"
        pagination:
          $ref: "#/components/schemas/Pagination"
    AdminUserDetail:
      type: object
      required: [user, player, goals, habits, items, activity]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        player:
          type: ["object", "null"]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            level:
              type: integer
            xp:
              type: integer
            hp:
              type: integer
            maxHp:
              type: integer
            gold:
              type: integer
            strength:
              type: integer
            endurance:
              type: integer
            dexterity:
              type: integer
            faith:
              type: integer
            avatarClass:
              type: string
            equippedSkills:
              type: array
              items:
                type: string
            createdAt:
              type: string
        goals:
          type: array
          items:
            type: object
        habits:
          type: array
          items:
            type: object
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
    AdminItemEntry:
      type: object
      required: [id, userId, name, description, rarity, type, obtainedAt, username]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rarity:
          type: string
        type:
          type: string
        obtainedAt:
          type: string
        username:
          type: string
    AdminItemList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminItemEntry"
        pagination:
          $ref: "#/components/schemas/Pagination"