# lib/api-zod/src/generated/types/adminBossInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoulForge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminBossInputDifficulty } from './adminBossInputDifficulty';

export interface AdminBossInput {
  name: string;
  description?: string;
  difficulty?: AdminBossInputDifficulty;
  maxHp?: number;
  xpReward?: number;
  goldReward?: number;
  lootTable?: string[];
  /** @nullable */
  imageUrl?: string | null;
  isCustom?: boolean;
}